china_yiwu/requirements.txt
aiogram==3.17.0
aiogram_dialog==2.3.1
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic_settings==2.7.1
aio-pika==9.5.4
faststream==0.5.34
loguru==0.7.3
aiosqlite==0.21.0
alembic==1.14.1
pytz==2025.1
apscheduler==3.11.0
fastapi==0.115.8
uvicorn==0.34.0
python-telegram-bot==22.0
python-dotenv==1.0.1
requests==2.32.3
psycopg2==2.9.9
Jinja2==3.1.6
openpyxl==3.1.5
pillow==11.2.1